composite actions (#42)
## READY

Modified process to make use of Composite GitHub Actions. This process
leaves the existing reusable workflows as a dependency for existing
"customers", only shifts to make use of the new Composite GitHub Action.

This allows us to decouple authentication tokens from the implementation
of Guardian.

This is a temporary solution to allow use to start using Guardian
internally. A more full-featured and performant action should be written
in the future.