add change detection to limit plans/applies (#62)
Adds the ability to detect git changes and only run guardian for those
directories.

A workflow job will be used to look for all terraform files with a
`backend ".*" {` syntax. If a file exists in a directory with this
syntax, it will be considered the root of a module and will be a target
for plan/apply.

A second step will look for what files have changed and create the
intersection between changed directories and the root directories.