feat: create run command to replicate guardian admin functionality (#183)
Closes #152 

Adds new functionality to the init command to validate github actor
permission levels for the repository. This is useful when a repository
administrator wants to create a more powerful workflow that can run
sensitive commands to administer terraform, but wants to restrict it to
repo admins or specific repo permission levels.

Adds a new run command that can run arbitrary terraform sub commands.
This also adds the ability to restrict which commands are allowed to be
run. This is useful when running in a github workflow for
workflow_dispatch so repository admins can create a workflow that can
only run a subset of commands for users to administer failures.

NOTE: all of this works and is secure because the workflow is run from
main, so only administrators can create workflows and modify and merge
with approval