feat: migrate plan-status to go (#192)
Migrate the existing plan-status action to the workflows command. This
command is used by a github workflow as a post status job. The users can
configured required status checks to ensure this check is required
because when using a matrix strategy, the plan job has dynamic names and
makes it impossible to have required status checks for all possible
directories.

This change also cleans up required flags for several commands and adds
tests for them.