feat: add destroy functionality (#349)
Adding the ability to detect removed directories and if a user supplied
meta comment of `GUARDIAN_DESTROY=terraform/backends/project1` exists,
the supplied directory will be planned and applied as a destroy instead.

This functionality only cover the `entrypoints` command to identify the
appropriate directories. The output of this command is a breaking
change, as the JSON output has additional fields to support the modified
and destroyed directories.

The follow up to this will be to implement the functionality in the plan
and apply commands to us the destroy flag.

NOTE: Some of the test data files just had a license header added and
can be ignored.

GUARDIAN_DESTROY=terraform/project2