Save plan file as JSON file in plan command (#394)
This change writes the plan to a JSON file, in addition to the existing
plan binary, so that it can be parsed by other commands (specifically
the `policy` command).

Added a new `output-dir` flag to specify the directory that plan binary
and json files are written to. This is helpful for writing to temp
directories in unit tests, so that any written files are cleaned up
after running unit tests.