fix: apply for destroy (#399)
When applying a destroy operation, we need a separate checkout action
via the github workflow so we cannot rely on metadata (which comes from
download the plan AFTER the entrypoints are resolved). Instead, we will
load the pull request body for the commit that was pushed for apply to
determine the same set of entrypoints.