Add a GitHub Organization's team memberships in policy fetch-data subcommand (#423)
This includes a refactoring of the team memberships query, which was
originally embedded inside `GetLatestApprovers`, into its own Platform
interface method, `GetTeamMemberships`. This is to expose the team
membership mappings of an organization in the result of the `guardian
policy fetch-data` command, under the `github.team_memberships` key, so
that the team memberships can be reused for writing policy definitions.