Add ListChangeRequestsByCommit skeleton to Platform interface (#466)
This method is required for Guardian Apply to report its status on a
change request. For GitHub and Guardian, the PR/MR ID will not be
available from the GitHub Action/GitLab Pipeline environment when
merging, so the ID needs to be looked up based on the SHA of the commit.