chore: update release process (#480)
The previously release process didnt work for release/** branches.

Instead of relying on creating a tag, the new process allows users to
generate a release preview in a pull request. Upon approval and push,
the release will be triggered.

This process is much cleaner than the previous process and mimics what
is done for jvs and pkg.

Decided to stick with goreleaser for now.