fix: fixes entrypoint comments in push (#507)
This was a miss in the v2 branch, but has since been fixed on main (v3).
When back porting new functionality to this branch, this bug was
discovered. This is mostly copy and paste between main and this branch
with some refactoring to missing pieces that exist on main.

- Refactor list pull requests for commit into function (this doesn't
need to be in the Platform interface because we aren't adding GitLab
functionality into v2, it will exist in v3)

- Fix bug by using the new function in `ReportStatus` and
`ReportEntrypointsSummary` to get pull request info when number is
missing